+++ src/commands/aircraft/fixinfo.ts
@@ -2,7 +2,7 @@ import { MessageCommandDefinition } from '../../lib/command';
 import { makeEmbed, makeLines } from '../../lib/embed';
 import { CommandCategory } from '../../constants';
 
-const FIX_INFO_URL = 'https://cdn.discordapp.com/attachments/630529130189946882/914959833709563905/nd-orl-fixinfo2.png';
+const FIX_INFO_URL = `${process.env.IMAGE_BASE_URL}/a32nx/fixinfo.png`;
 
 const fixInfoEmbed = makeEmbed({
     title: 'FlyByWire A32NX | Fix Info',