+++ src/commands/general/navRouteTypes.ts
@@ -2,7 +2,7 @@ import { MessageCommandDefinition } from '../../lib/command';
 import { makeEmbed } from '../../lib/embed';
 import { CommandCategory } from '../../constants';
 
-const navRoutesURL = 'https://media.discordapp.net/attachments/838062729398976522/1051934378269364244/38c571b663cd3204e89707b223a5b7de.png';
+const navRoutesURL = `${process.env.IMAGE_BASE_URL}general/navroutes.png`;
 
 const genericNavRouteTypesEmbed = makeEmbed({
     title: 'FlyByWire | Navigation Route Types',