+++ src/commands/moderation/faq.ts
@@ -2,7 +2,7 @@ import { CommandDefinition } from '../../lib/command';
 import { makeEmbed, makeLines } from '../../lib/embed';
 import { Channels, CommandCategory, Roles, RoleGroups } from '../../constants';
 
-const FLIGHT_DECK_IMAGE_URL = 'https://media.discordapp.net/attachments/838062729398976522/889484404697743381/unknown.png?width=1214&height=683';
+const FLIGHT_DECK_IMAGE_URL = `${process.env.IMAGE_BASE_URL}moderation/faq.png`;
 
 const faqEmbeds = [
     makeEmbed({