+++ src/commands/support/screenshot.ts
@@ -2,7 +2,7 @@ import { MessageCommandDefinition } from '../../lib/command';
 import { makeEmbed, makeLines } from '../../lib/embed';
 import { CommandCategory } from '../../constants';
 
-const SCREENSHOT_HELP_URL = 'https://docs.flybywiresim.com/fbw-a32nx/assets/support-guide/cockpit-screenshot.jpg';
+const SCREENSHOT_HELP_URL = `${process.env.IMAGE_BASE_URL}support/screenshot.jpg`;
 
 const screenshotEmbed = makeEmbed({
     title: 'FlyByWire Support | How to take a good screenshot',